Add config opt. to mute exceptions on unseen text
This addition allows the user to allow top2vec to still
do keyword searches even if all the keywords are not
in the models vocab.

The 'CONFIG_IGNORE_UNSEEN_WORDS' environment
variable, when set to 1, will enable filtering on the
keywords to prevent a ValueError exception when
the words don't exist in the model's vocab.